Для установки скорости порта более 57600 был изменён тип данных с uint16_t на uint62_t. Реализована инициализация модуля E18. Частично реализована работа по разбору потока данных из модуля E18. Небольшой пересмотр кода.

+++ drivers/MtmZigbee.h
@@ -40,11 +40,11 @@
 
 void *mtmZigbeeDeviceThread(void *device);
 
-int32_t mtmZigbeeInit(int32_t mode, uint8_t *path, uint32_t speed);
+int32_t mtmZigbeeInit(int32_t mode, uint8_t *path, uint64_t speed);
 
 void mtmZigbeePktListener(DBase *dBase, int32_t threadId);
 
-speed_t mtmZigbeeGetSpeed(uint32_t speed);
+speed_t mtmZigbeeGetSpeed(uint64_t speed);
 
 bool mtmZigbeeGetRun();
